+++ src/lib.rs
@@ -10,6 +10,9 @@ mod handler;
 
 pub mod prelude;
 
+#[cfg(feature = "actix-web")]
+pub mod web;
+
 pub use actor::AsyncContextExt;
 pub use arbiter::ArbiterExt;
 pub use handler::ResponseStdFuture;