+++ src/prelude.rs
@@ -1,2 +1,4 @@
 pub use crate::{actor::AsyncContextExt, arbiter::ArbiterExt};
 
+#[cfg(feature = "actix-web")]
+pub use crate::web::RouteExt;