+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 
 - Update to be compatible with `futures 0.3.0-alpha.13`.
 - Increase the minimum requirement of rustc to `1.35.0-nightly`.
+- New feature: "actix-web"
+  - Add `RouteExt::with_std_async` and `RouteExt::with_std_async_config`, like `Route::with_async` for async functions.
 
 0.2.0 (October 23, 2018)
 ------------------------